requirements.txt
django
djangorestframework
gunicorn
django-heroku